+++ KernelBench/level1/62_conv_standard_2D__square_input__asymmetric_kernel.py
@@ -17,7 +17,7 @@ class Model(nn.Module):
     """
     def __init__(self, in_channels: int, out_channels: int, kernel_size: tuple, stride: int = 1, padding: int = 0, dilation: int = 1, groups: int = 1, bias: bool = False):
         super(Model, self).__init__()
-        self.conv2d = nn.Conv2d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, dilation=dilation, groups=groups, bias=bias, dtype=torch.bfloat16)
+        self.conv2d = nn.Conv2d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, dilation=dilation, groups=groups, bias=bias, dtype=torch.half)
         
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
@@ -40,7 +40,7 @@ width = 512
 height = 512
 
 def get_inputs():
-    x = torch.rand(batch_size, in_channels, height, width, dtype=torch.bfloat16)
+    x = torch.rand(batch_size, in_channels, height, width, dtype=torch.half)
     return [x]
 
 def get_init_inputs():
