+++ KernelBench/level1/69_conv_transposed_2D__asymmetric_input__asymmetric_kernel.py
@@ -18,7 +18,7 @@ class Model(nn.Module):
     """
     def __init__(self, in_channels: int, out_channels: int, kernel_size: tuple, stride: tuple = (1, 1), padding: tuple = (0, 0), output_padding: tuple = (0, 0), dilation: tuple = (1, 1), groups: int = 1, bias: bool = False):
         super(Model, self).__init__()
-        self.conv_transpose2d = nn.ConvTranspose2d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, output_padding=output_padding, dilation=dilation, groups=groups, bias=bias, dtype=torch.bfloat16)
+        self.conv_transpose2d = nn.ConvTranspose2d(in_channels, out_channels, kernel_size, stride=stride, padding=padding, output_padding=output_padding, dilation=dilation, groups=groups, bias=bias, dtype=torch.half)
         
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
@@ -41,7 +41,7 @@ height_in = 128
 width_in = 256
 
 def get_inputs():
-    x = torch.rand(batch_size, in_channels, height_in, width_in, dtype=torch.bfloat16)
+    x = torch.rand(batch_size, in_channels, height_in, width_in, dtype=torch.half)
     return [x]
 
 def get_init_inputs():
