+++ KernelBench/level1/93_masked_cumsum.py
@@ -29,7 +29,7 @@ input_shape = (32768,)
 dim = 1
 
 def get_inputs():
-    x = torch.rand(batch_size, *input_shape, dtype=torch.bfloat16)
+    x = torch.rand(batch_size, *input_shape, dtype=torch.half)
     mask = torch.randint(0, 2, x.shape).bool()  # Random boolean mask
     return [x, mask]
