+++ KernelBench/level4/20_facebook-bart-large_bs32_seq256.py
@@ -7,7 +7,7 @@ class Model(torch.nn.Module):
         super().__init__()
         self.model_name = model_name
         self.config = config
-        self.model = AutoModelForCausalLM.from_pretrained(self.model_name, config=self.config, dtype=torch.bfloat16)
+        self.model = AutoModelForCausalLM.from_pretrained(self.model_name, config=self.config, dtype=torch.half)
 
     def forward(self, x):
         return self.model(x).logits
